chore(playground): move components to separate files
